+++ src/Makefile
@@ -1,5 +1,5 @@
 all:
-	gcc inquisitor.c packetManage.c -o a.out -lpcap -Wno-deprecated-declarations && ./a.out
+	gcc inquisitor.c packetManage.c utils.c -o a.out -lpcap -Wno-deprecated-declarations && ./a.out
 
 d:
-	gcc inquisitor.c packetManage.c -o a.out -I/System/Volumes/Data/sgoinfre/goinfre/Perso/$USER/homebrew/Cellar/libnet/1.2/include -lpcap -v -Wall -g && ./a.out
+	gcc inquisitor.c packetManage.c utils.c -o a.out -I/System/Volumes/Data/sgoinfre/goinfre/Perso/$USER/homebrew/Cellar/libnet/1.2/include -lpcap -v -Wall -g && ./a.out
