+++ src/packetManage.h
@@ -1,6 +1,32 @@
 #ifndef PKT_MNG 
 #define PKT_MNG
 
+typedef struct TCP_packet{
+
+    struct pcap_pkthdr* packet_header;
+    struct ether_header* eth_header;
+
+    //all the headers in a tcp
+
+     u_char* packet;
+     u_char* ethernet_header;
+     u_char *ip_header; // packet + ethernet_header_length;     const u_char *tcp_header;
+     u_char *tcp_header; // packet + ethernet_header_length;     const u_char *tcp_header;
+     u_char *payload;
+     u_char protocol; //in this program we are going to use only FTP and ARP
+     u_char *port;
+
+    // Longitud de las cabeceras (bytes)
+    int ethernet_header_length;    // Constante
+    int ip_header_length;           // = 9
+    int tcp_header_length;          //=tcp_header_length * 4;
+    int payload_length;             //= 
+    int total_headers_size;         // = ethernet_header_length+ip_header_length+tcp_header_length;
+                                    //
+
+
+}TCP_packet;
+
 
 /**
  * @brief Encuentra la carga (payload) de un paquete.
@@ -11,18 +37,6 @@
  */
 void my_packet_handler(u_char*, const struct pcap_pkthdr*, const u_char*);
 
-/**
- * @brief Mostrar información de un paquete.
- * 
- * @param[in] pcap_packet   Paquete.
- */
-void print_packet_info(const u_char*, struct pcap_pkthdr);
-
-
-typedef struct tcp_packet{
-
-
-}tcp_packet;
 
 
 #endif
